+++ components/Layout.js
@@ -18,15 +18,18 @@ class Layout extends React.PureComponent {
       <div>
         <Head>
           <script src="ace/src-min/ace.js" />
-          <link rel="stylesheet" type="text/css" href="markup-editor/lib/codemirror.css" />
           <link rel="stylesheet" type="text/css" href="hljs/styles/github.css" />
           <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/github-markdown-css/2.9.0/github-markdown.min.css" />
+          <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />
         </Head>
         <Link href="/">
           <button>Home</button>
         </Link>
         <Link href="/markdown">
           <button>Markdown</button>
+        </Link>
+        <Link href="/asciidoc">
+          <button>Asciidoc</button>
         </Link>
         <Link href="/embedding">
           <button>Embed</button>