+++ next.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  webpack: config => Object.assign(config, { node: { fs: 'empty' } }),
+};